+++ src/components/Admin/ParentAdmin.jsx
@@ -35,6 +35,7 @@ import {
 } from '../../utils/imageFallback';
 import GotifySettings from './GotifySettings';
 import HomeAssistantSettings from './HomeAssistantSettings';
+import NextcloudSettings from './NextcloudSettings';
 import ProblemReportsPanel from './ProblemReportsPanel';
 import WebPushSettings from './WebPushSettings';
 
@@ -511,6 +512,7 @@ export default function ParentAdmin({ onOpenFamilyTree }) {
           </div>
         </section>
 
+        <NextcloudSettings />
         <WebPushSettings />
         <GotifySettings />
         <HomeAssistantSettings />